+++ CHANGELOG.md
@@ -1,3 +1,5 @@
+# 0.0.1.rc1
+
 # 0.0.1.beta1
 
 * [FEATURE] basic remotes support